+++ test/test-create.ps1
@@ -1,5 +1,5 @@
 #Requires -RunAsAdministrator
-#Requires -Version 7.5
+#requires -version 7
 
 <#
 .SYNOPSIS